+++ lib/animation_wrappers.dart
@@ -1,8 +1,20 @@
 library animation_wrappers;
 
-///animation wrappers
-export 'package:animation_wrappers/Animations/faded_scale_animation.dart';
-export 'package:animation_wrappers/Animations/faded_slide_animation.dart';
+/// animation wrappers
+
+/// FadedScaleAnimation
+export 'package:animation_wrappers/animations/faded_scale_animation.dart';
+
+/// FadedSlideAnimation
+export 'package:animation_wrappers/animations/faded_slide_animation.dart';
+
+/// FadeAnimation
+export 'package:animation_wrappers/animations/fade_animation.dart';
+
+/// ScaleAnimation
+export 'package:animation_wrappers/animations/scale_animation.dart';
 
 ///animated bottom bars
-export 'package:animation_wrappers/AnimatedBottomNavigationBars/animated_text_bottom_bar.dart';
+
+/// AnimatedTextBottomBar
+export 'package:animation_wrappers/animated_bottom_navigation_bars/animated_text_bottom_bar.dart';